backend/requirements.txt
Django==1.11.5
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
psycopg2==2.7.4
PyJWT==1.6.4
pytz==2018.4
